refactor: rename to .ts & refactor indents in news file
